test: ensure bpflogsize tests can handle large logs
With recent kernel versions, the output from the BPF log can be too
much for the test to handle in reasonable time, and the test fails
with a timeout.  Since the test is only looking for data to match
near the end of the log, use tail to focus on just that.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>